requirements.txt
requests
flask
redis
hiredis
python-rapidjson
